Fix gcc commandline
Neither flags nor pflags is able to work with the different gcc
commandline arguments.
Therefore clean the commandline from flags we don't care about
before passing it to pflags. Pflag is used as it allows to collect
multiple flags of the same type in a slice.

This is done as follows:
- Remove the string arguments by looking at the next argument to not
leave remains that would later be detected as input files.
- Remove bool flags
- Add a singe dash to some arguments that need to pass pflag.

After wiping the unnecessary parts we can pass the clean arguments to
pflags.