Cleanup master & db code
* master main
Remove all related functionality to a separate package named 'master'
and introduced 2 flags for setting rpc addr & port and db addr & port

* pkg master
This contains the implementation of gRPC server. This part of the code
'speeks' protobuf. No other part of the application should use the pb
messages directly. This package should take care of extracting all
information needed and transform that to other structs

* pkg database
This package manages data from and in to db, also should take care of
other db related functionality. So this layer speacks golang structs
that represents our data in the database