Sync database writing
Enable concurrent builds and circumvent dgraphs missing unique term.
The AddNode function recursively inserts a node's derivedFrom nodes and
the node itself to a queue (channel). This channel is processed by a
goroutine that checks the node to be ready -- all derived nodes are in
the database and therefore have a uid before adding it to the database.

This way we can handle builds with more than one job and do not stall
build.