Fix address issue
Due to setting the AddressOptional in init() it was always true and
therefore useless and broke all commands that needed an address.

Remove the conceptually wrong AddressOptional boolean and setup the
address in the setup function. Whenever a command does not need to call
the server it just does not call out to it.