Sanitize nodes
AlterFileNode and AlterPackageNode are not used anymore.
The alteration of the filenode or the package node is taking place in the GetAllInfoData.
No need to sanitize the fields of the package node as we just called it from the database but we want to
sanitize the additional info of the package node.