Change nodeType to per node predicate
Remove nodeType that needs to be set correctly to a predicate named like
the actual message. This makes it easier to query and filter by that
predicate without looking up the corresponding int to the nodeType. Also
this is less error prone because the actual value assigned is
irrelevant.

Clean up database code

To have the node predicates serialized when writing to the database a
dummy value is inserted automatically.