Loop should not break if we have not found the same hash
After adding the extra missing pieces nodes, while we search for the wanted targets to
add them to the package node we might want to add an altered target with different hash.
In that case if we parse the old path target we want to continue the loop to find the
new target with the matching hash to be added to the package nodes.