Extend connect and disconnect commands (#268)
* Extend connect cmd for the case of a project node

Extend connect cmd to connect packages to a project node.

* Check which type of nodes we are trying to connect to our node

A user may try to add a wrong type to a node. For instance,
try to add file nodes to a project node.

* Extend disconnect cmd for the case of a project node

Extend disconnect cmd to disconnect package nodes from a project node.

* Remove switch in disconnect cmd

No need for a switch. Cleaner code.

* Rename edge -> predicates

We are actually removing predicates.

* Remove switch

Switch with one case is not worth a switch.

* Use a more proper name for updating the project node