GetPackageNode returns a stream of packages
Since the database structure changed, we may have more than one package
node in our database. GetPackageΝode returns a stream of packages.
If we want to get a specific package node from the database, we can
call the function parameterised by a package with the desired name.
The function will return a stream with only the desired node.
Otherwise it will return a stream with all the packages in the database.