Fix dgraph queries
If we want to query for a filenode with a specific path, we can't query
it dirrectly, as the path is no longer a field of filenode but it is a field
in pathInfo. If we try to use a filter fot the path in the query, the query
would return all the filenodes in the db and just one path connected to the
filenode we were searching. The block would be huge and would exceed the available
result size.
A better solution is to use the path to query for the hash and then with the hash
we can dirrectly get the specific node we need.