Choose to loop recursively a query with or without a specified depth
When we are quering for instance for a hash in a big project and this filenode is
derived from many other files, if we use the "@recurse(loop: false)", then the query
will ignore paths which lead to a loop while traversing. In our case that means that
the query will ignore information that we need.