Add custom template for scancode output
Currently, scancode analyzer uses the default json output from scancode.
This requires further processing/parsing on the go side to bring the
data in the format required by the protobuf spec. The main issue with
that is that it's done by iterating through the parsed file list for
each fileNode received from the master service.
Using the path as a key in the custom structure we can reduce the
computational complexity by using a map-key access with the fileNode's
path.
Also it allows us to create a go structure that conforms both to
the custom output and the protobuf specs. Most of the parsing is then no
longer required.